Atualização do projeto

+++ src/service/JogadorService.java
@@ -1,5 +1,6 @@
 package service;
 
+import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.util.List;
 
@@ -8,12 +9,23 @@ import org.json.JSONObject;
 import org.simpleframework.http.Query;
 import org.simpleframework.http.Request;
 
+import dao.ContadorCartaoDAO;
+import dao.EstatisticaDAO;
+import dao.FichaMedicaDAO;
 import dao.JogadorDAO;
+import dao.TimeDAO;
+import model.Contadordecartoes;
+import model.Estatistica;
+import model.FichaMedica;
 import model.Jogador;
 import model.enums.Escala;
 
 public class JogadorService {
 	private JogadorDAO jogadorDAO = new JogadorDAO();
+	private FichaMedicaDAO fichaDAO = new FichaMedicaDAO();
+	private ContadorCartaoDAO contadorDAO = new ContadorCartaoDAO();
+	private EstatisticaDAO estatisticaDAO = new EstatisticaDAO();
+	private TimeDAO timeDAO = new TimeDAO();
 
 	public String consultarJogador(Integer id, Request request) {
 		try {
@@ -22,7 +34,7 @@ public class JogadorService {
 
 		} catch (Exception e) {
 			e.printStackTrace();
-			return "Erro ao consultar Jogador";
+			return "<ERRO>Erro ao consultar Jogador";
 		}
 	}
 
@@ -31,7 +43,7 @@ public class JogadorService {
 			return listaJogadoresJSON().toString();
 		} catch (Exception e) {
 			e.printStackTrace();
-			return "Erro ao consultar Jogadores";
+			return "<ERRO>Erro ao consultar Jogadores";
 		}
 	}
 
@@ -51,7 +63,7 @@ public class JogadorService {
 			return query.get("url");
 		} catch (Exception e) {
 			e.printStackTrace();
-			return "Erro ao adicionar Jogador";
+			return "<ERRO>Erro ao adicionar Jogador";
 		}
 	}
 
@@ -72,7 +84,7 @@ public class JogadorService {
 			return query.get("url");
 		} catch (Exception e) {
 			e.printStackTrace();
-			return "Erro ao atualizar Jogador";
+			return "<ERRO>Erro ao atualizar Jogador";
 		}
 	}
 
@@ -83,12 +95,16 @@ public class JogadorService {
 			Jogador jogador = new Jogador();
 			jogador.setId(Integer.parseInt(query.get("id")));
 			
-			jogadorDAO.delete(jogador);
+			if(possuiVinculo(jogador)) {
+				return "<ERRO>O jogador não pode ser excluído porque possui vínculo com outros registros do sistema.";
+			} else {
+				jogadorDAO.delete(jogador);
+			}
 			
 			return query.get("url");
 		} catch (Exception e) {
 			e.printStackTrace();
-			return "Erro ao excluir Jogador.";
+			return "<ERRO>Erro ao excluir Jogador.";
 		}
 	}
 
@@ -103,5 +119,27 @@ public class JogadorService {
 		obj.put("listaJogadores", new JSONArray(listaJogadores));
 
 		return obj;
+	}
+	
+	private boolean possuiVinculo(Jogador j) throws NumberFormatException, FileNotFoundException, IOException {
+		boolean possui = false;
+		
+		for(FichaMedica f : fichaDAO.getAll()) {
+			possui = possui || f.getJogador().equals(j);
+		}
+		
+		for(Contadordecartoes c : contadorDAO.getAll()) {
+			possui = possui || c.getJogador().equals(j);
+		}
+		
+		for(Estatistica e : estatisticaDAO.getAll()) {
+			possui = possui || e.getJogador().equals(j);
+		}
+		
+		for(Jogador jogadorTime : timeDAO.get().getListaJogadores()) {
+			possui = possui || jogadorTime.equals(j);
+		}
+		
+		return possui;
 	}
 }
